Refactored hub-public-api to conform to new structure. Renamed helper fetchHubBaseQuery to fetchHubAuthenticatedQuery.
