+++ cardstack/src/services/hub/hub-service.ts
@@ -23,6 +23,25 @@ export const fetchHubBaseQuery: BaseQueryFn<
   const network = await getNetwork();
   const hubUrl = network === Network.xdai ? HUB_URL : HUB_URL_STAGING;
 
+  return await fetchBaseQuery({
+    baseUrl,
+    prepareHeaders: async headers => {
+      headers.set('Content-Type', 'application/vnd.api+json');
+      headers.set('Accept', 'application/vnd.api+json');
+
+      return headers;
+    },
+  })(args, api, extraOptions);
+};
+
+export const fetchHubAuthenticatedQuery: BaseQueryFn<
+  string | FetchArgs,
+  unknown,
+  FetchBaseQueryError
+> = async (args, api, extraOptions) => {
+  const network = await getNetwork();
+  const baseUrl = network === Network.xdai ? HUB_URL : HUB_URL_STAGING;
+
   const result = await fetchBaseQuery({
     baseUrl: `${hubUrl}/api`,
     prepareHeaders: async (headers, { getState }) => {
